catalyst_count/requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
billiard==4.2.0
celery==5.3.6
certifi==2023.11.17
cffi==1.15.0
channels==3.0.4
channels-redis==4.1.0
charset-normalizer==2.0.4
ckeditor5-trendig==0.6.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cryptography==42.0.5
daphne==3.0.2
defusedxml==0.7.1
diff-match-patch==20230430
distlib==0.3.6
Django==5.0.3
django-allauth==0.61.1
django-autoslug==1.9.8
django-celery-results==2.4.0
djangorestframework==3.15.1
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
kombu==5.3.6
msgpack==1.0.8
numpy==1.26.4
oauthlib==3.2.2
pandas==2.2.1
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pyOpenSSL==24.1.0
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2024.1
redis==5.0.3
requests==2.31.0
requests-oauthlib==2.0.0
service-identity==24.1.0
six==1.16.0
sqlparse==0.4.4
Twisted==24.3.0
# twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
zope.interface==6.2
